refactor: split transfer tests to multiple files
This is to allow better granularity on choosing the tests in the ci-min-smoke-test.js
